# backend/server/api_activity.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A ActivityApiController binds http requests to an api service and writes the service results to the http response
type ActivityApiController struct {
	service ActivityApiServicer
}

// NewActivityApiController creates a default api controller
func NewActivityApiController(s ActivityApiServicer) Router {
	return &ActivityApiController{service: s}
}

// Routes returns all of the api route for the ActivityApiController
func (c *ActivityApiController) Routes() Routes {
	return Routes{
		{
			"GetSexActivityByDate",
			strings.ToUpper("Get"),
			"/sexActivity/{userId}",
			c.GetSexActivityByDate,
		},
	}
}

// GetSexActivityByDate -
func (c *ActivityApiController) GetSexActivityByDate(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	userId := params["userId"]
	date := query.Get("date")
	result, err := c.service.GetSexActivityByDate(r.Context(), userId, date)
	//If an error occured, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	//If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}