# backend/server/api_bleeding.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A BleedingApiController binds http requests to an api service and writes the service results to the http response
type BleedingApiController struct {
	service BleedingApiServicer
}

// NewBleedingApiController creates a default api controller
func NewBleedingApiController(s BleedingApiServicer) Router {
	return &BleedingApiController{service: s}
}

// Routes returns all of the api route for the BleedingApiController
func (c *BleedingApiController) Routes() Routes {
	return Routes{
		{
			"GetBleedingByDate",
			strings.ToUpper("Get"),
			"/bleeding/{userId}",
			c.GetBleedingByDate,
		},
	}
}

// GetBleedingByDate -
func (c *BleedingApiController) GetBleedingByDate(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	userId := params["userId"]
	date := query.Get("date")
	result, err := c.service.GetBleedingByDate(r.Context(), userId, date)
	//If an error occured, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	//If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}