# backend/server/api_desire.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A DesireApiController binds http requests to an api service and writes the service results to the http response
type DesireApiController struct {
	service DesireApiServicer
}

// NewDesireApiController creates a default api controller
func NewDesireApiController(s DesireApiServicer) Router {
	return &DesireApiController{service: s}
}

// Routes returns all of the api route for the DesireApiController
func (c *DesireApiController) Routes() Routes {
	return Routes{
		{
			"GetSexDesireByDate",
			strings.ToUpper("Get"),
			"/sexDesire/{userId}",
			c.GetSexDesireByDate,
		},
	}
}

// GetSexDesireByDate -
func (c *DesireApiController) GetSexDesireByDate(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	userId := params["userId"]
	date := query.Get("date")
	result, err := c.service.GetSexDesireByDate(r.Context(), userId, date)
	//If an error occured, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	//If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}