# backend/server/api_mood.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A MoodApiController binds http requests to an api service and writes the service results to the http response
type MoodApiController struct {
	service MoodApiServicer
}

// NewMoodApiController creates a default api controller
func NewMoodApiController(s MoodApiServicer) Router {
	return &MoodApiController{service: s}
}

// Routes returns all of the api route for the MoodApiController
func (c *MoodApiController) Routes() Routes {
	return Routes{
		{
			"GetMoodByDate",
			strings.ToUpper("Get"),
			"/mood/{userId}",
			c.GetMoodByDate,
		},
	}
}

// GetMoodByDate -
func (c *MoodApiController) GetMoodByDate(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	userId := params["userId"]
	date := query.Get("date")
	result, err := c.service.GetMoodByDate(r.Context(), userId, date)
	//If an error occured, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	//If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}