# backend/server/api_pain.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A PainApiController binds http requests to an api service and writes the service results to the http response
type PainApiController struct {
	service PainApiServicer
}

// NewPainApiController creates a default api controller
func NewPainApiController(s PainApiServicer) Router {
	return &PainApiController{service: s}
}

// Routes returns all of the api route for the PainApiController
func (c *PainApiController) Routes() Routes {
	return Routes{
		{
			"GetPainByDate",
			strings.ToUpper("Get"),
			"/pain/{userId}",
			c.GetPainByDate,
		},
	}
}

// GetPainByDate -
func (c *PainApiController) GetPainByDate(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	userId := params["userId"]
	date := query.Get("date")
	result, err := c.service.GetPainByDate(r.Context(), userId, date)
	//If an error occured, encode the error with the status code
	if err != nil {
		EncodeJSONResponse(err.Error(), &result.Code, w)
		return
	}
	//If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}