# backend/server/model_bleeding.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Bleeding - Bleeding information for a day
type Bleeding struct {

	// UUID
	Id string `json:"id,omitempty"`

	Strength string `json:"strength,omitempty"`

	Exclude string `json:"exclude,omitempty"`
}