# backend/server/model_cervical_mucus.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// CervicalMucus - Cervical Mucus information for a day
type CervicalMucus struct {

	// UUID
	Id string `json:"id,omitempty"`

	Feeling string `json:"feeling,omitempty"`

	Texture string `json:"texture,omitempty"`

	Exclude string `json:"exclude,omitempty"`
}