# backend/server/model_cervix.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Cervix - Cervix information for a day
type Cervix struct {

	// UUID
	Id string `json:"id,omitempty"`

	Opening string `json:"opening,omitempty"`

	Firmness string `json:"firmness,omitempty"`

	Position string `json:"position,omitempty"`

	Excluce string `json:"excluce,omitempty"`
}