# backend/server/model_day.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Day - A Day entry
type Day struct {

	// UUID
	Id string `json:"id,omitempty"`

	// Day key used to encrypt sensitive information for the day.
	DayKey string `json:"day_key,omitempty"`

	Version string `json:"version,omitempty"`

	Date string `json:"date,omitempty"`

	Bleeding Bleeding `json:"bleeding,omitempty"`

	Cervical CervicalMucus `json:"cervical,omitempty"`

	Cervix Cervix `json:"cervix,omitempty"`

	SexActivity SexualActivity `json:"sexActivity,omitempty"`

	SexDesire SexualDesire `json:"sexDesire,omitempty"`

	Pain Pain `json:"pain,omitempty"`

	Mood Mood `json:"mood,omitempty"`
}