# backend/server/model_mood.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Mood - Mood information for a day
type Mood struct {

	// UUID
	Id string `json:"id,omitempty"`

	Type string `json:"type,omitempty"`
}