# backend/server/model_pain.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Pain - Pain information for a day
type Pain struct {

	// UUID
	Id string `json:"id,omitempty"`

	Type string `json:"type,omitempty"`
}