# backend/server/model_sexual_activity.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// SexualActivity - Sexual activity information for a day
type SexualActivity struct {

	// UUID
	Id string `json:"id,omitempty"`

	Type string `json:"type,omitempty"`

	Contraceptives string `json:"contraceptives,omitempty"`
}