# backend/server/model_sexual_desire.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// SexualDesire - Sexual desire information for a day
type SexualDesire struct {

	// UUID
	Id string `json:"id,omitempty"`

	Intensity string `json:"intensity,omitempty"`
}