# backend/server/model_temperature.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// Temperature - Temperature information for a day
type Temperature struct {

	// UUID
	Id string `json:"id,omitempty"`

	Value string `json:"value,omitempty"`

	Timestamp string `json:"timestamp,omitempty"`

	Note string `json:"note,omitempty"`

	Exclude string `json:"exclude,omitempty"`
}