# backend/server/model_user.go
/*
 * Cyla App API
 *
 * API for the period tracking app 'Cyla'
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package server

// User - An User of the app
type User struct {

	// UUID
	Id string `json:"id,omitempty"`

	UserKeyBackup string `json:"user_key_backup,omitempty"`
}